pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rackbrain"
version = "1.0.0"
description = "Automated Jira ticket triage using YAML rules"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "PyYAML>=6.0.1,<7.0",
  "requests>=2.25.1,<3.0",
  "pymysql>=1.0.2,<2.0",
  "urllib3>=1.26.0,<2.0",
]

[project.scripts]
rackbrain = "rackbrain.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rackbrain*"]

[tool.setuptools.package-data]
rackbrain = ["py.typed"]
